+++ R9_407_V1.1/Drivers/BSP/Common/common.h
@@ -37,7 +37,7 @@ typedef struct {
 } AverageFilter;
 
 
-double Value_limit(double min_value ,double current_value ,double max_value);
+int32_t Value_limit(int32_t min_value ,int32_t current_value ,int32_t max_value);
 int32_t  Value_Resetzero(int32_t min_value ,int32_t current_value ,int32_t max_value);
 int32_t slopelimitx( int32_t value,int32_t increvalue) ;   
 int32_t slopelimity( int32_t value,int32_t increvalue) ;
